+++ src/kernel/memory/enable_paging.s
@@ -1,5 +1,6 @@
 .section .text
 .global enable_paging
+.global switch_pd
 
 enable_paging:
 	mov 4(%esp), %eax // get page directory address
@@ -10,3 +11,7 @@ enable_paging:
 	or $0x80000001, %eax
 	mov %eax, %cr0
 	ret
+
+switch_pd: // switch page directory
+	mov 4(%esp), %eax
+	mov %eax, %cr3
